pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brain-tumor-xai"
version = "0.1.0"
description = "Reusable baseline for brain MRI classification with XAI"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
dependencies = [
  "captum>=0.7,<1.0",
  "matplotlib>=3.8,<4.0",
  "numpy>=1.26,<3.0",
  "pillow>=10.0,<12.0",
  "scikit-learn>=1.4,<2.0",
  "torch>=2.2,<3.0",
  "torchvision>=0.17,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[tool.setuptools.packages.find]
where = ["src"]
